requirements.txt
annotated-types==0.6.0
anyio==4.0.0
APScheduler==3.10.4
astor==0.8.1
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==2.0.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==46.0.3
decorator==5.1.1
exceptiongroup==1.2.0
fastapi==0.110.1
greenlet==3.1.1
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
idna==3.4
imbalanced-learn==0.13.0
isort==6.0.1
joblib==1.4.2
lxml==4.9.3
numpy==1.26.0
opt-einsum==3.3.0
packaging==24.0
pandas==2.2.3
Pillow==10.0.1
protobuf==4.24.3
pycparser==2.23
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.6.1
scipy==1.15.1
six==1.16.0
sklearn-compat==0.1.3
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.34
sqlmodel==0.0.25
starlette==0.37.2
threadpoolctl==3.5.0
thulac==0.2.2
typing_extensions==4.15.0
tzdata==2025.1
tzlocal==5.2
urllib3==2.0.5
uvicorn==0.29.0
watchfiles==0.21.0
webencodings==0.5.1
websockets==12.0
